Looking for a better way to get the files, when it gets uploaded
